pyproject.toml
[project]
name = "reddit-agent-manus"
version = "0.1.0"
description = ""
authors = [
    {name = "ZHOU HUIWEI",email = "you@example.com"}
]
requires-python = ">=3.14"
dependencies = [
    "pydantic (>=2.12.5,<3.0.0)",
    "tenacity (>=9.1.4,<10.0.0)",
    "rich (>=14.3.3,<15.0.0)",
    "openai (>=1.0.0,<2.0.0)",
    "docker (>=7.0.0,<8.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "anthropic (>=0.1.0,<1.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
# We only use Poetry for dependency management here.
# Disable installing the current project as a package to avoid
# "No file/folder found for package reddit-agent-manus".
package-mode = false
